+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 #### Features
 
 * Embedded [Maputnik](https://github.com/maputnik/editor) style editor
+* Service Info page with viewer code snippets 
 
 <a name="0.7.0"></a>
 ## 0.7.0 (2017-03-12)